+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - PayPal donate button
 - setpgrp flag
 
+### Fixed
+- Reading from closed channel now fails only when all data has been retrieved
+
 ### Removed
 - Python 3.4 support